// touristagency-front/generated/models/HotelIn.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tourist agency
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ServiceClass } from './ServiceClass';
import {
    ServiceClassFromJSON,
    ServiceClassFromJSONTyped,
    ServiceClassToJSON,
} from './ServiceClass';

/**
 * 
 * @export
 * @interface HotelIn
 */
export interface HotelIn {
    /**
     * 
     * @type {string}
     * @memberof HotelIn
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof HotelIn
     */
    address: string;
    /**
     * 
     * @type {ServiceClass}
     * @memberof HotelIn
     */
    serviceClass: ServiceClass;
    /**
     * 
     * @type {boolean}
     * @memberof HotelIn
     */
    isGuideIncluded: boolean;
}

/**
 * Check if a given object implements the HotelIn interface.
 */
export function instanceOfHotelIn(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "address" in value;
    isInstance = isInstance && "serviceClass" in value;
    isInstance = isInstance && "isGuideIncluded" in value;

    return isInstance;
}

export function HotelInFromJSON(json: any): HotelIn {
    return HotelInFromJSONTyped(json, false);
}

export function HotelInFromJSONTyped(json: any, ignoreDiscriminator: boolean): HotelIn {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'address': json['address'],
        'serviceClass': ServiceClassFromJSON(json['serviceClass']),
        'isGuideIncluded': json['isGuideIncluded'],
    };
}

export function HotelInToJSON(value?: HotelIn | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'address': value.address,
        'serviceClass': ServiceClassToJSON(value.serviceClass),
        'isGuideIncluded': value.isGuideIncluded,
    };
}